requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dill==0.3.3
dnspython==2.0.0
fastapi==0.63.0
future==0.18.2
h11==0.12.0
idna==2.10
mongoengine==0.22.1
pydantic==1.7.3
pyfunctional==1.4.2
pymongo==3.11.2
requests==2.25.1
six==1.15.0
soupsieve==2.1
starlette==0.13.6
tabulate==0.8.7
urllib3==1.26.2
uvicorn==0.13.3
